Update: Removed helper file
